Update testFileForHello-World.txt
The auto text above is "Update testFileforHello-World.txt.   Hoping to see if @GrayWolf-19 gets this and can comment on it.   Then maybe we can do other tests where we each apply changes.   @GrayWolf-19 please comment on this change and whether or not you are able to see it and or work with it.